requirements.txt
beautifulsoup4
requests
flask
flask_sqlalchemy
flask-login
gunicorn
python-dotenv
Flask-WTF
Flask-Limiter
Flask-APScheduler
